+++ leeger/calculator/all_time_calculator/ScoringShareAllTimeCalculator.py
@@ -2,6 +2,7 @@ from typing import Optional
 
 from leeger.calculator.all_time_calculator.PointsScoredAllTimeCalculator import PointsScoredAllTimeCalculator
 from leeger.calculator.parent.AllTimeCalculator import AllTimeCalculator
+from leeger.calculator.year_calculator import ScoringShareYearCalculator
 from leeger.decorator.validators import validateLeague
 from leeger.model.league.League import League
 from leeger.util.Deci import Deci
@@ -86,3 +87,107 @@ class ScoringShareAllTimeCalculator(AllTimeCalculator):
                                                                    ownerId] / totalPointsScoredInLeague) * Deci("100")
 
         return ownerIdAndOpponentScoringShare
+
+    @classmethod
+    @validateLeague
+    def getMaxScoringShare(cls, league: League, **kwargs) -> dict[str, Optional[Deci]]:
+        """
+        Returns the Max Scoring Share for each Owner in the given League.
+        Returns None for an Owner if they have no games played in the range.
+
+        Example response:
+            {
+            "someOwnerId": Deci("10.7"),
+            "someOtherOwnerId": Deci("14.2"),
+            "yetAnotherOwnerId": Deci("12.1"),
+            ...
+            }
+        """
+
+        ownerIdAndMaxScoringShare = dict()
+
+        allOwnerIds = LeagueNavigator.getAllOwnerIds(league)
+
+        for ownerId in allOwnerIds:
+            ownerIdAndMaxScoringShare[ownerId] = None
+
+        maxScoringSharesByYearTeamIds = cls._getAllResultDictsByYear(league,
+                                                                     ScoringShareYearCalculator.getMaxScoringShare,
+                                                                     **kwargs)
+        # swap out team IDs for owner IDs
+        maxScoringSharesByYear = dict()
+        for yearNumber, maxScoringSharesByTeamId in maxScoringSharesByYearTeamIds.items():
+            maxScoringSharesByYear[yearNumber] = dict()
+            for teamId, maxScoringShare in maxScoringSharesByTeamId.items():
+                ownerId = LeagueNavigator.getTeamById(league, teamId).ownerId
+                maxScoringSharesByYear[yearNumber][ownerId] = maxScoringShare
+
+        ownerIdAndMaxScoringShares: dict[str, list] = dict()
+        for yearNumber in maxScoringSharesByYear.keys():
+            for ownerId in allOwnerIds:
+                if ownerId in ownerIdAndMaxScoringShares.keys() and ownerIdAndMaxScoringShares[ownerId] is not None:
+                    ownerIdAndMaxScoringShares[ownerId].append(maxScoringSharesByYear[yearNumber][ownerId])
+                else:
+                    ownerIdAndMaxScoringShares[ownerId] = [maxScoringSharesByYear[yearNumber][ownerId]]
+
+        for ownerId in allOwnerIds:
+            # remove all None values from list
+            ownerIdAndMaxScoringShares[ownerId] = [i for i in ownerIdAndMaxScoringShares[ownerId] if i is not None]
+            if len(ownerIdAndMaxScoringShares[ownerId]) > 0:
+                ownerIdAndMaxScoringShare[ownerId] = max(ownerIdAndMaxScoringShares[ownerId])
+            else:
+                ownerIdAndMaxScoringShare[ownerId] = None
+
+        return ownerIdAndMaxScoringShare
+
+    @classmethod
+    @validateLeague
+    def getMinScoringShare(cls, league: League, **kwargs) -> dict[str, Optional[Deci]]:
+        """
+        Returns the Min Scoring Share for each Owner in the given League.
+        Returns None for an Owner if they have no games played in the range.
+
+        Example response:
+            {
+            "someOwnerId": Deci("10.7"),
+            "someOtherOwnerId": Deci("14.2"),
+            "yetAnotherOwnerId": Deci("12.1"),
+            ...
+            }
+        """
+
+        ownerIdAndMinScoringShare = dict()
+
+        allOwnerIds = LeagueNavigator.getAllOwnerIds(league)
+
+        for ownerId in allOwnerIds:
+            ownerIdAndMinScoringShare[ownerId] = None
+
+        minScoringSharesByYearTeamIds = cls._getAllResultDictsByYear(league,
+                                                                     ScoringShareYearCalculator.getMinScoringShare,
+                                                                     **kwargs)
+        # swap out team IDs for owner IDs
+        minScoringSharesByYear = dict()
+        for yearNumber, minScoringSharesByTeamId in minScoringSharesByYearTeamIds.items():
+            minScoringSharesByYear[yearNumber] = dict()
+            for teamId, maxScoringShare in minScoringSharesByTeamId.items():
+                ownerId = LeagueNavigator.getTeamById(league, teamId).ownerId
+                minScoringSharesByYear[yearNumber][ownerId] = maxScoringShare
+
+        ownerIdAndMinScoringShares: dict[str, list] = dict()
+        for yearNumber in minScoringSharesByYear.keys():
+            for ownerId in allOwnerIds:
+                if ownerId in ownerIdAndMinScoringShares.keys() and ownerIdAndMinScoringShares[ownerId] is not None:
+                    ownerIdAndMinScoringShares[ownerId].append(minScoringSharesByYear[yearNumber][ownerId])
+                else:
+                    ownerIdAndMinScoringShares[ownerId] = [minScoringSharesByYear[yearNumber][ownerId]]
+
+        for ownerId in allOwnerIds:
+            # remove all None values from list
+            ownerIdAndMinScoringShares[ownerId] = [i for i in ownerIdAndMinScoringShares[ownerId] if i is not None]
+            if len(ownerIdAndMinScoringShares[ownerId]) > 0:
+                ownerIdAndMinScoringShare[ownerId] = min(ownerIdAndMinScoringShares[ownerId])
+            else:
+                ownerIdAndMinScoringShare[ownerId] = None
+
+        return ownerIdAndMinScoringShare