+++ leeger/model/stat/YearStatSheet.py
@@ -37,6 +37,8 @@ class YearStatSheet:
     # Scoring Share
     scoringShare: dict[str, Deci]
     opponentScoringShare: dict[str, Deci]
+    maxScoringShare: dict[str, Deci]
+    minScoringShare: dict[str, Deci]
 
     # Single Score
     maxScore: dict[str, float | int]
@@ -84,6 +86,8 @@ class YearStatSheet:
             ("Opponent Points Scored Per Game", self.opponentPointsScoredPerGame),
             ("Scoring Share", self.scoringShare),
             ("Opponent Scoring Share", self.opponentScoringShare),
+            ("Max Scoring Share", self.maxScoringShare),
+            ("Min Scoring Share", self.minScoringShare),
             ("Max Score", self.maxScore),
             ("Min Score", self.minScore),
             ("Scoring Standard Deviation", self.scoringStandardDeviation),